Refactor the KafkaClusterOperations to include managing the zookeeper resources.
It turns out it's simpler to move CompositeOperations and delete the
ZookeeperClusterOperations than to deletete to the ZookeeperClusterOperations
