+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operations/cluster/KafkaClusterOperations.java
@@ -9,8 +9,11 @@ import io.strimzi.controller.cluster.operations.resource.StatefulSetOperations;
 import io.strimzi.controller.cluster.resources.ClusterDiffResult;
 import io.strimzi.controller.cluster.resources.KafkaCluster;
 import io.strimzi.controller.cluster.resources.Storage;
+import io.strimzi.controller.cluster.resources.ZookeeperCluster;
+import io.vertx.core.AsyncResult;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
+import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -49,7 +52,13 @@ public class KafkaClusterOperations extends AbstractClusterOperations<KafkaClust
         this.pvcOperations = pvcOperations;
     }
 
-    private final CompositeOperation<KafkaCluster> create = new CompositeOperation<KafkaCluster>() {
+    public void create(String namespace, String name, Handler<AsyncResult<Void>> handler) {
+        // TODO Don't pass the same handler
+        execute(namespace, name, createZk, ar -> {});
+        execute(namespace, name, createKafka, handler);
+    }
+
+    private final CompositeOperation<KafkaCluster> createKafka = new CompositeOperation<KafkaCluster>() {
 
         @Override
         public ClusterOperation<KafkaCluster> getCluster(String namespace, String name){
@@ -76,7 +85,37 @@ public class KafkaClusterOperations extends AbstractClusterOperations<KafkaClust
         }
     };
 
-    private final CompositeOperation<KafkaCluster> delete = new CompositeOperation<KafkaCluster>() {
+
+    private final CompositeOperation<ZookeeperCluster> createZk = new CompositeOperation<ZookeeperCluster>() {
+
+        @Override
+        public ClusterOperation<ZookeeperCluster> getCluster(String namespace, String name) {
+            return new ClusterOperation<>(ZookeeperCluster.fromConfigMap(configMapOperations.get(namespace, name)), null);
+        }
+
+        @Override
+        public Future<?> composite(String namespace, ClusterOperation<ZookeeperCluster> clusterOp) {
+            ZookeeperCluster zk = clusterOp.cluster();
+            List<Future> result = new ArrayList<>(4);
+
+            if (zk.isMetricsEnabled()) {
+                result.add(configMapOperations.create(zk.generateMetricsConfigMap()));
+            }
+
+            result.add(serviceOperations.create(zk.generateService()));
+
+            result.add(serviceOperations.create(zk.generateHeadlessService()));
+
+            result.add(statefulSetOperations.create(zk.generateStatefulSet(isOpenShift)));
+
+            return CompositeFuture.join(result);
+        }
+    };
+
+
+
+
+    private final CompositeOperation<KafkaCluster> deleteKafka = new CompositeOperation<KafkaCluster>() {
         @Override
         public Future<?> composite(String namespace, ClusterOperation<KafkaCluster> clusterOp) {
             KafkaCluster kafka = clusterOp.cluster();
@@ -110,22 +149,67 @@ public class KafkaClusterOperations extends AbstractClusterOperations<KafkaClust
         }
     };
 
+
+    private final CompositeOperation<ZookeeperCluster> deleteZk = new CompositeOperation<ZookeeperCluster>() {
+        @Override
+        public Future<?> composite(String namespace, ClusterOperation<ZookeeperCluster> clusterOp) {
+            ZookeeperCluster zk = clusterOp.cluster();
+            boolean deleteClaims = zk.getStorage().type() == Storage.StorageType.PERSISTENT_CLAIM
+                    && zk.getStorage().isDeleteClaim();
+            List<Future> result = new ArrayList<>(4 + (deleteClaims ? zk.getReplicas() : 0));
+
+            // start deleting configMap operation only if metrics are enabled,
+            // otherwise the future is already complete (for the "join")
+            if (zk.isMetricsEnabled()) {
+                result.add(configMapOperations.delete(namespace, zk.getMetricsConfigName()));
+            }
+
+            result.add(serviceOperations.delete(namespace, zk.getName()));
+
+            result.add(serviceOperations.delete(namespace, zk.getHeadlessName()));
+
+            result.add(statefulSetOperations.delete(namespace, zk.getName()));
+
+
+            if (deleteClaims) {
+                for (int i = 0; i < zk.getReplicas(); i++) {
+                    result.add(pvcOperations.delete(namespace, zk.getVolumeName() + "-" + zk.getName() + "-" + i));
+                }
+            }
+
+            return CompositeFuture.join(result);
+        }
+
+        @Override
+        public ClusterOperation<ZookeeperCluster> getCluster(String namespace, String name) {
+            StatefulSet ss = statefulSetOperations.get(namespace, ZookeeperCluster.zookeeperClusterName(name));
+            return new ClusterOperation<ZookeeperCluster>(ZookeeperCluster.fromStatefulSet(ss, namespace, name), null);
+        }
+    };
+
+    @Override
+    public void delete(String namespace, String name, Handler<AsyncResult<Void>> handler) {
+        // TODO don't pass the same handler
+        execute(namespace, name, deleteKafka, ar -> {});
+        execute(namespace, name, deleteZk, handler);
+    }
+
     @Override
     protected CompositeOperation<KafkaCluster> createOp() {
-        return create;
+        return createKafka;
     }
 
     @Override
     protected CompositeOperation<KafkaCluster> deleteOp() {
-        return delete;
+        return deleteKafka;
     }
 
     @Override
     protected CompositeOperation<KafkaCluster> updateOp() {
-        return update;
+        return updateKafka;
     }
 
-    private final CompositeOperation<KafkaCluster> update = new CompositeOperation<KafkaCluster>() {
+    private final CompositeOperation<KafkaCluster> updateKafka = new CompositeOperation<KafkaCluster>() {
         @Override
         public Future<?> composite(String namespace, ClusterOperation<KafkaCluster> clusterOp) {
             KafkaCluster kafka = clusterOp.cluster();
@@ -161,79 +245,197 @@ public class KafkaClusterOperations extends AbstractClusterOperations<KafkaClust
             }
             return new ClusterOperation<>(kafka, diff);
         }
+
+
+        private Future<Void> scaleDown(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
+            if (diff.isScaleDown())    {
+                log.info("Scaling down stateful set {} in namespace {}", kafka.getName(), namespace);
+                return statefulSetOperations.scaleDown(namespace, kafka.getName(), kafka.getReplicas());
+            }
+            else {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> patchService(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
+            if (diff.isDifferent()) {
+                return serviceOperations.patch(namespace, kafka.getName(), kafka.patchService(serviceOperations.get(namespace, kafka.getName())));
+            }
+            else
+            {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> patchHeadlessService(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
+            if (diff.isDifferent()) {
+                return serviceOperations.patch(namespace, kafka.getHeadlessName(),
+                        kafka.patchHeadlessService(serviceOperations.get(namespace, kafka.getHeadlessName())));
+            }
+            else
+            {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> patchStatefulSet(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
+            if (diff.isDifferent()) {
+                return statefulSetOperations.patch(namespace, kafka.getName(), false,
+                        kafka.patchStatefulSet(statefulSetOperations.get(namespace, kafka.getName())));
+            } else {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> patchMetricsConfigMap(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
+            if (diff.isMetricsChanged()) {
+                return configMapOperations.patch(namespace, kafka.getMetricsConfigName(),
+                        kafka.patchMetricsConfigMap(configMapOperations.get(namespace, kafka.getMetricsConfigName())));
+            } else {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> rollingUpdate(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
+            Future<Void> rollingUpdate = Future.future();
+
+            if (diff.isRollingUpdate()) {
+                statefulSetOperations.rollingUpdate(namespace, kafka.getName(),
+                        rollingUpdate.completer());
+            }
+            else {
+                rollingUpdate.complete();
+            }
+
+            return rollingUpdate;
+        }
+
+        private Future<Void> scaleUp(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
+            if (diff.isScaleUp()) {
+                return statefulSetOperations.scaleUp(namespace, kafka.getName(), kafka.getReplicas());
+            }
+            else {
+                return Future.succeededFuture();
+            }
+        }
     };
 
-    private Future<Void> scaleDown(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
-        if (diff.isScaleDown())    {
-            log.info("Scaling down stateful set {} in namespace {}", kafka.getName(), namespace);
-            return statefulSetOperations.scaleDown(namespace, kafka.getName(), kafka.getReplicas());
+    private final CompositeOperation<ZookeeperCluster> updateZk = new CompositeOperation<ZookeeperCluster>() {
+        @Override
+        public Future<?> composite(String namespace, ClusterOperation<ZookeeperCluster> operation) {
+            ZookeeperCluster zk = operation.cluster();
+            ClusterDiffResult diff = operation.diff();
+            Future<Void> chainFuture = Future.future();
+
+            scaleDown(zk, namespace, diff)
+                    .compose(i -> patchService(zk, namespace, diff))
+                    .compose(i -> patchHeadlessService(zk, namespace, diff))
+                    .compose(i -> patchStatefulSet(zk, namespace, diff))
+                    .compose(i -> patchMetricsConfigMap(zk, namespace, diff))
+                    .compose(i -> rollingUpdate(zk, namespace, diff))
+                    .compose(i -> scaleUp(zk, namespace, diff))
+                    .compose(chainFuture::complete, chainFuture);
+
+            return chainFuture;
         }
-        else {
-            return Future.succeededFuture();
+
+        private Future<Void> scaleDown(ZookeeperCluster zk, String namespace, ClusterDiffResult diff) {
+            if (diff.isScaleDown())    {
+                log.info("Scaling down stateful set {} in namespace {}", zk.getName(), namespace);
+                return statefulSetOperations.scaleDown(namespace, zk.getName(), zk.getReplicas());
+            }
+            else {
+                return Future.succeededFuture();
+            }
         }
+
+        private Future<Void> patchService(ZookeeperCluster zk, String namespace, ClusterDiffResult diff) {
+            if (diff.isDifferent()) {
+                return serviceOperations.patch(namespace, zk.getName(),
+                        zk.patchService(serviceOperations.get(namespace, zk.getName())));
+            }
+            else
+            {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> patchHeadlessService(ZookeeperCluster zk, String namespace, ClusterDiffResult diff) {
+            if (diff.isDifferent()) {
+                return serviceOperations.patch(namespace, zk.getHeadlessName(),
+                        zk.patchHeadlessService(serviceOperations.get(namespace, zk.getHeadlessName())));
+            }
+            else
+            {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> patchStatefulSet(ZookeeperCluster zk, String namespace, ClusterDiffResult diff) {
+            if (diff.isDifferent()) {
+                return statefulSetOperations.patch(namespace, zk.getName(), false,
+                        zk.patchStatefulSet(statefulSetOperations.get(namespace, zk.getName())));
+            }
+            else
+            {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> patchMetricsConfigMap(ZookeeperCluster zk, String namespace, ClusterDiffResult diff) {
+            if (diff.isMetricsChanged()) {
+                return configMapOperations.patch(namespace, zk.getMetricsConfigName(),
+                        zk.patchMetricsConfigMap(configMapOperations.get(namespace, zk.getMetricsConfigName())));
+            } else {
+                return Future.succeededFuture();
+            }
+        }
+
+        private Future<Void> rollingUpdate(ZookeeperCluster zk, String namespace, ClusterDiffResult diff) {
+            Future<Void> rollingUpdate = Future.future();
+
+            if (diff.isRollingUpdate()) {
+                statefulSetOperations.rollingUpdate(namespace, zk.getName(),
+                        rollingUpdate.completer());
+            }
+            else {
+                rollingUpdate.complete();
+            }
+
+            return rollingUpdate;
+        }
+
+        private Future<Void> scaleUp(ZookeeperCluster zk, String namespace, ClusterDiffResult diff) {
+            if (diff.isScaleUp()) {
+                return statefulSetOperations.scaleUp(namespace, zk.getName(), zk.getReplicas());
+            }
+            else {
+                return Future.succeededFuture();
+            }
+        }
+
+        @Override
+        public ClusterOperation<ZookeeperCluster> getCluster(String namespace, String name) {
+            ClusterDiffResult diff;
+            ZookeeperCluster zk;
+            ConfigMap zkConfigMap = configMapOperations.get(namespace, name);
+
+            if (zkConfigMap != null)    {
+                zk = ZookeeperCluster.fromConfigMap(zkConfigMap);
+                log.info("Updating Zookeeper cluster {} in namespace {}", zk.getName(), namespace);
+                diff = zk.diff(configMapOperations, statefulSetOperations, namespace);
+            } else {
+                throw new IllegalStateException("ConfigMap " + name + " doesn't exist anymore in namespace " + namespace);
+            }
+
+            return new ClusterOperation<>(zk, diff);
+        }
+    };
+
+    public void update(String namespace, String name, Handler<AsyncResult<Void>> handler) {
+        // TODO don't pass the same handler
+        execute(namespace, name, updateZk, ar -> {});
+        execute(namespace, name, updateKafka, handler);
     }
 
-    private Future<Void> patchService(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
-        if (diff.isDifferent()) {
-            return serviceOperations.patch(namespace, kafka.getName(), kafka.patchService(serviceOperations.get(namespace, kafka.getName())));
-        }
-        else
-        {
-            return Future.succeededFuture();
-        }
-    }
-
-    private Future<Void> patchHeadlessService(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
-        if (diff.isDifferent()) {
-            return serviceOperations.patch(namespace, kafka.getHeadlessName(),
-                    kafka.patchHeadlessService(serviceOperations.get(namespace, kafka.getHeadlessName())));
-        }
-        else
-        {
-            return Future.succeededFuture();
-        }
-    }
-
-    private Future<Void> patchStatefulSet(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
-        if (diff.isDifferent()) {
-            return statefulSetOperations.patch(namespace, kafka.getName(), false,
-                    kafka.patchStatefulSet(statefulSetOperations.get(namespace, kafka.getName())));
-        }
-        else
-        {
-            return Future.succeededFuture();
-        }
-    }
-
-    private Future<Void> patchMetricsConfigMap(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
-        if (diff.isMetricsChanged()) {
-            return configMapOperations.patch(namespace, kafka.getMetricsConfigName(),
-                    kafka.patchMetricsConfigMap(configMapOperations.get(namespace, kafka.getMetricsConfigName())));
-        } else {
-            return Future.succeededFuture();
-        }
-    }
-
-    private Future<Void> rollingUpdate(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
-        Future<Void> rollingUpdate = Future.future();
-
-        if (diff.isRollingUpdate()) {
-            statefulSetOperations.rollingUpdate(namespace, kafka.getName(),
-                    rollingUpdate.completer());
-        }
-        else {
-            rollingUpdate.complete();
-        }
-
-        return rollingUpdate;
-    }
-
-    private Future<Void> scaleUp(KafkaCluster kafka, String namespace, ClusterDiffResult diff) {
-        if (diff.isScaleUp()) {
-            return statefulSetOperations.scaleUp(namespace, kafka.getName(), kafka.getReplicas());
-        }
-        else {
-            return Future.succeededFuture();
-        }
-    }
 }
