+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/resources/ZookeeperCluster.java
@@ -43,12 +43,12 @@ public class ZookeeperCluster extends AbstractCluster {
     // N/A
 
     // Configuration keys
-    private static String KEY_IMAGE = "zookeeper-image";
-    private static String KEY_REPLICAS = "zookeeper-nodes";
-    private static String KEY_HEALTHCHECK_DELAY = "zookeeper-healthcheck-delay";
-    private static String KEY_HEALTHCHECK_TIMEOUT = "zookeeper-healthcheck-timeout";
-    private static String KEY_METRICS_CONFIG = "zookeeper-metrics-config";
-    private static String KEY_STORAGE = "zookeeper-storage";
+    public static String KEY_IMAGE = "zookeeper-image";
+    public static String KEY_REPLICAS = "zookeeper-nodes";
+    public static String KEY_HEALTHCHECK_DELAY = "zookeeper-healthcheck-delay";
+    public static String KEY_HEALTHCHECK_TIMEOUT = "zookeeper-healthcheck-timeout";
+    public static String KEY_METRICS_CONFIG = "zookeeper-metrics-config";
+    public static String KEY_STORAGE = "zookeeper-storage";
 
     // Zookeeper configuration keys
     private static String KEY_ZOOKEEPER_NODE_COUNT = "ZOOKEEPER_NODE_COUNT";
